server/requirements.txt
flask>=2.2.2
flask-cors>=3.0.10
PyJWT>=2.3.0
pymongo>=4.4.2
gunicorn>=20.1.0
requests>=2.27.1
Werkzeug>=3.0.0